+++ src/plugins/platforms/xcb/qxcbscreen.h
@@ -226,7 +226,7 @@ private:
     QRect m_availableGeometry;
     Qt::ScreenOrientation m_orientation = Qt::PrimaryOrientation;
     QXcbCursor *m_cursor;
-    int m_refreshRate = 60;
+    qreal m_refreshRate = 60.0;
     int m_pixelDensity = 1;
     QEdidParser m_edid;
 };
